- remove team_id from being created in model_has_roles in default permission migration - add check for team_id column in migration where we are trying to add it again

+++ database/migrations/2023_01_05_011210_make_team_id_nullable_on_model_has_roles_table.php
@@ -4,14 +4,15 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     public function up()
     {
-        Schema::table('model_has_roles', function (Blueprint $table) {
-            $table->dropPrimary('team_id');
-            $table->dropColumn('team_id');
-        });
+        if (Schema::hasColumn('model_has_roles', 'team_id')) {
+            Schema::table('model_has_roles', function (Blueprint $table) {
+                $table->dropPrimary('team_id');
+                $table->dropColumn('team_id');
+            });
+        }
         Schema::table('model_has_roles', function (Blueprint $table) {
             $table->unsignedBigInteger('team_id')->nullable()->index();
         });
